KTOR-1689 Extract ContentConverter to shared module
