+++ gradle/compatibility.gradle
@@ -27,6 +27,7 @@ apiValidation {
         project(":ktor-io"),
         project(":ktor-server"),
         project(":ktor-features"),
+        project(":ktor-shared"),
     ]
 
     while (!queue.isEmpty()) {
