+++ ktor-client/ktor-client-features/ktor-client-json/ktor-client-gson/build.gradle.kts
@@ -11,7 +11,7 @@ kotlin.sourceSets {
         dependencies {
             api(project(":ktor-client:ktor-client-cio"))
             api(project(":ktor-client:ktor-client-features:ktor-client-json:ktor-client-json-tests"))
-            api(project(":ktor-features:ktor-gson"))
+            api(project(":ktor-shared:ktor-shared-serialization-gson"))
         }
     }
 }
