+++ ktor-client/ktor-client-features/ktor-client-json/ktor-client-jackson/build.gradle.kts
@@ -14,7 +14,7 @@ kotlin.sourceSets {
         dependencies {
             api(project(":ktor-client:ktor-client-cio"))
             api(project(":ktor-client:ktor-client-features:ktor-client-json:ktor-client-json-tests"))
-            api(project(":ktor-features:ktor-gson"))
+            api(project(":ktor-shared:ktor-shared-serialization-gson"))
         }
     }
 }
