+++ ktor-client/ktor-client-tests/build.gradle.kts
@@ -67,7 +67,7 @@ kotlin.sourceSets {
             api(project(":ktor-server:ktor-server-jetty"))
             api(project(":ktor-features:ktor-auth"))
             api(project(":ktor-features:ktor-websockets"))
-            api(project(":ktor-features:ktor-serialization"))
+            api(project(":ktor-shared:ktor-shared-serialization-kotlinx"))
             api("ch.qos.logback:logback-classic:$logback_version")
             api("junit:junit:$junit_version")
             api("org.jetbrains.kotlin:kotlin-test-junit:$kotlin_version")
