--- ktor-features/ktor-gson/api/ktor-gson.api
@@ -1,13 +1,0 @@
-public final class io/ktor/gson/GsonConverter : io/ktor/features/ContentConverter {
-	public fun <init> ()V
-	public fun <init> (Lcom/google/gson/Gson;)V
-	public synthetic fun <init> (Lcom/google/gson/Gson;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
-	public fun convertForReceive (Lio/ktor/util/pipeline/PipelineContext;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-	public fun convertForSend (Lio/ktor/util/pipeline/PipelineContext;Lio/ktor/http/ContentType;Ljava/lang/Object;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-}
-
-public final class io/ktor/gson/GsonSupportKt {
-	public static final fun gson (Lio/ktor/features/ContentNegotiation$Configuration;Lio/ktor/http/ContentType;Lkotlin/jvm/functions/Function1;)V
-	public static synthetic fun gson$default (Lio/ktor/features/ContentNegotiation$Configuration;Lio/ktor/http/ContentType;Lkotlin/jvm/functions/Function1;ILjava/lang/Object;)V
-}
-
