--- ktor-features/ktor-jackson/api/ktor-jackson.api
@@ -1,13 +1,0 @@
-public final class io/ktor/jackson/JacksonConverter : io/ktor/features/ContentConverter {
-	public fun <init> ()V
-	public fun <init> (Lcom/fasterxml/jackson/databind/ObjectMapper;)V
-	public synthetic fun <init> (Lcom/fasterxml/jackson/databind/ObjectMapper;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
-	public fun convertForReceive (Lio/ktor/util/pipeline/PipelineContext;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-	public fun convertForSend (Lio/ktor/util/pipeline/PipelineContext;Lio/ktor/http/ContentType;Ljava/lang/Object;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-}
-
-public final class io/ktor/jackson/JacksonConverterKt {
-	public static final fun jackson (Lio/ktor/features/ContentNegotiation$Configuration;Lio/ktor/http/ContentType;Lkotlin/jvm/functions/Function1;)V
-	public static synthetic fun jackson$default (Lio/ktor/features/ContentNegotiation$Configuration;Lio/ktor/http/ContentType;Lkotlin/jvm/functions/Function1;ILjava/lang/Object;)V
-}
-
