--- ktor-features/ktor-jackson/build.gradle
@@ -1,11 +1,0 @@
-description = ''
-kotlin {
-    sourceSets {
-        jvmMain {
-            dependencies {
-                api group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: jackson_version
-                api group: 'com.fasterxml.jackson.module', name: 'jackson-module-kotlin', version: jackson_kotlin_version
-            }
-        }
-    }
-}
