--- ktor-features/ktor-serialization/api/ktor-serialization.api
@@ -1,34 +1,0 @@
-public final class io/ktor/serialization/ConverterSupportKt {
-	public static final fun serialization (Lio/ktor/features/ContentNegotiation$Configuration;Lio/ktor/http/ContentType;Lkotlinx/serialization/BinaryFormat;)V
-	public static final fun serialization (Lio/ktor/features/ContentNegotiation$Configuration;Lio/ktor/http/ContentType;Lkotlinx/serialization/StringFormat;)V
-}
-
-public final class io/ktor/serialization/JsonSupportKt {
-	public static final fun getDefaultJson ()Lkotlinx/serialization/json/Json;
-	public static final fun getDefaultJsonConfiguration ()Lkotlinx/serialization/json/Json;
-	public static final fun json (Lio/ktor/features/ContentNegotiation$Configuration;Lkotlinx/serialization/json/Json;Lio/ktor/http/ContentType;)V
-	public static final fun json (Lio/ktor/features/ContentNegotiation$Configuration;Lkotlinx/serialization/json/Json;Lkotlinx/serialization/modules/SerializersModule;Lio/ktor/http/ContentType;)V
-	public static synthetic fun json$default (Lio/ktor/features/ContentNegotiation$Configuration;Lkotlinx/serialization/json/Json;Lio/ktor/http/ContentType;ILjava/lang/Object;)V
-	public static synthetic fun json$default (Lio/ktor/features/ContentNegotiation$Configuration;Lkotlinx/serialization/json/Json;Lkotlinx/serialization/modules/SerializersModule;Lio/ktor/http/ContentType;ILjava/lang/Object;)V
-	public static final fun serialization (Lio/ktor/features/ContentNegotiation$Configuration;)V
-	public static final fun serialization (Lio/ktor/features/ContentNegotiation$Configuration;Lio/ktor/http/ContentType;)V
-	public static final fun serialization (Lio/ktor/features/ContentNegotiation$Configuration;Lio/ktor/http/ContentType;Lkotlinx/serialization/json/Json;)V
-}
-
-public final class io/ktor/serialization/SerializationConverter : io/ktor/features/ContentConverter {
-	public synthetic fun <init> ()V
-	public fun <init> (Lkotlinx/serialization/BinaryFormat;)V
-	public fun <init> (Lkotlinx/serialization/StringFormat;Ljava/nio/charset/Charset;)V
-	public synthetic fun <init> (Lkotlinx/serialization/StringFormat;Ljava/nio/charset/Charset;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
-	public synthetic fun <init> (Lkotlinx/serialization/json/Json;)V
-	public synthetic fun <init> (Lkotlinx/serialization/json/Json;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
-	public fun convertForReceive (Lio/ktor/util/pipeline/PipelineContext;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-	public fun convertForSend (Lio/ktor/util/pipeline/PipelineContext;Lio/ktor/http/ContentType;Ljava/lang/Object;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-}
-
-public final class io/ktor/serialization/SerializationConverterKt {
-	public static final fun SerializationConverter ()Lio/ktor/serialization/SerializationConverter;
-	public static final synthetic fun serialization (Lio/ktor/features/ContentNegotiation$Configuration;Lio/ktor/http/ContentType;Lkotlinx/serialization/json/Json;)V
-	public static synthetic fun serialization$default (Lio/ktor/features/ContentNegotiation$Configuration;Lio/ktor/http/ContentType;Lkotlinx/serialization/json/Json;ILjava/lang/Object;)V
-}
-
