--- ktor-features/ktor-serialization/build.gradle
@@ -1,10 +1,0 @@
-/*
- * Copyright 2014-2019 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
- */
-
-apply plugin: "kotlinx-serialization"
-
-kotlin.sourceSets.jvmMain.dependencies {
-    api "org.jetbrains.kotlinx:kotlinx-serialization-json:$serialization_version"
-}
-
