+++ ktor-features/ktor-websockets/jvm/test/io/ktor/tests/websocket/WebSocketWithContentNegotiationTest.kt
@@ -7,11 +7,14 @@ package io.ktor.tests.websocket
 import io.ktor.application.*
 import io.ktor.features.*
 import io.ktor.http.*
-import io.ktor.http.cio.websocket.Frame
-import io.ktor.request.*
+import io.ktor.http.cio.websocket.*
+import io.ktor.http.content.*
 import io.ktor.routing.*
 import io.ktor.server.testing.*
-import io.ktor.util.pipeline.*
+import io.ktor.shared.serialization.*
+import io.ktor.util.reflect.*
+import io.ktor.utils.io.*
+import io.ktor.utils.io.charsets.*
 import io.ktor.websocket.*
 import kotlin.test.*
 
@@ -22,15 +25,14 @@ class WebSocketWithContentNegotiationTest {
         application.install(WebSockets)
         application.install(ContentNegotiation) {
             val converter = object : ContentConverter {
-                override suspend fun convertForSend(
-                    context: PipelineContext<Any, ApplicationCall>,
+                override suspend fun serialize(
                     contentType: ContentType,
+                    charset: Charset,
+                    typeInfo: TypeInfo,
                     value: Any
-                ): Any? = fail("convertForSend shouldn't be invoked")
+                ): OutgoingContent? = fail("convertForSend shouldn't be invoked")
 
-                override suspend fun convertForReceive(
-                    context: PipelineContext<ApplicationReceiveRequest, ApplicationCall>
-                ): Any? {
+                override suspend fun deserialize(charset: Charset, typeInfo: TypeInfo, content: ByteReadChannel): Any? {
                     fail("convertForReceive shouldn't be invoked")
                 }
             }
