+++ ktor-shared/build.gradle
@@ -1,0 +1,9 @@
+description = 'shared functionality for client and server'
+
+subprojects {
+    kotlin.sourceSets {
+        commonMain.dependencies {
+            api project(':ktor-http')
+        }
+    }
+}
