+++ ktor-shared/ktor-shared-serialization-gson/jvm/src/GsonConverter.kt
@@ -1,0 +1,72 @@
+/*
+* Copyright 2014-2021 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+*/
+
+package io.ktor.shared.serializaion.gson
+
+import com.google.gson.*
+import io.ktor.http.*
+import io.ktor.http.content.*
+import io.ktor.shared.serialization.*
+import io.ktor.util.pipeline.*
+import io.ktor.util.reflect.*
+import io.ktor.utils.io.*
+import io.ktor.utils.io.charsets.*
+import io.ktor.utils.io.jvm.javaio.*
+import kotlinx.coroutines.*
+import kotlin.coroutines.*
+import kotlin.reflect.*
+import kotlin.reflect.jvm.*
+
+/**
+ * GSON converter for [ContentNegotiation] feature
+ */
+public class GsonConverter(private val gson: Gson = Gson()) : ContentConverter {
+
+    override suspend fun serialize(
+        contentType: ContentType,
+        charset: Charset,
+        typeInfo: TypeInfo,
+        value: Any
+    ): OutgoingContent? {
+        return TextContent(gson.toJson(value), contentType.withCharset(charset))
+    }
+
+    override suspend fun deserialize(charset: Charset, typeInfo: TypeInfo, content: ByteReadChannel): Any? {
+        if (gson.isExcluded(typeInfo.type)) {
+            throw ExcludedTypeGsonException(typeInfo.type)
+        }
+
+        return withContext(Dispatchers.IO) {
+            val reader = content.toInputStream().reader(charset)
+            gson.fromJson(reader, typeInfo.reifiedType)
+        }
+    }
+}
+
+private fun Gson.isExcluded(type: KClass<*>) =
+    excluder().excludeClass(type.java, false)
+
+@OptIn(ExperimentalCoroutinesApi::class)
+internal class ExcludedTypeGsonException(
+    private val type: KClass<*>
+) : Exception("Type ${type.jvmName} is excluded so couldn't be used in receive"),
+    CopyableThrowable<ExcludedTypeGsonException> {
+
+    override fun createCopy(): ExcludedTypeGsonException = ExcludedTypeGsonException(type).also {
+        it.initCause(this)
+    }
+}
+
+/**
+ * Register Gson to [ContentNegotiation] feature
+ */
+public fun Configuration.gson(
+    contentType: ContentType = ContentType.Application.Json,
+    block: GsonBuilder.() -> Unit = {}
+) {
+    val builder = GsonBuilder()
+    builder.apply(block)
+    val converter = GsonConverter(builder.create())
+    register(contentType, converter)
+}
