+++ ktor-shared/ktor-shared-serialization-jackson/jvm/src/JacksonConverter.kt
@@ -1,0 +1,74 @@
+/*
+* Copyright 2014-2021 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+*/
+
+package io.ktor.shared.serializaion.jackson
+
+import com.fasterxml.jackson.core.util.*
+import com.fasterxml.jackson.databind.*
+import com.fasterxml.jackson.module.kotlin.*
+import io.ktor.http.*
+import io.ktor.http.content.*
+import io.ktor.shared.serialization.*
+import io.ktor.util.reflect.*
+import io.ktor.utils.io.*
+import io.ktor.utils.io.charsets.*
+import io.ktor.utils.io.jvm.javaio.*
+import kotlinx.coroutines.*
+import kotlin.reflect.*
+import kotlin.reflect.jvm.*
+import kotlin.text.Charsets
+
+public class JacksonConverter(private val objectmapper: ObjectMapper = jacksonObjectMapper()) : ContentConverter {
+
+    override suspend fun serialize(
+        contentType: ContentType,
+        charset: Charset,
+        typeInfo: TypeInfo,
+        value: Any
+    ): OutgoingContent {
+        return OutputStreamContent(
+            {
+                if (charset == Charsets.UTF_8) {
+                    /*
+                    Jackson internally does special casing on UTF-8, presumably for performance reasons. Thus we pass an
+                    InputStream instead of a writer to let Jackson do it's thing.
+                     */
+                    objectmapper.writeValue(this, value)
+                } else {
+                    objectmapper.writeValue(this.writer(charset = charset), value)
+                }
+            },
+            contentType.withCharset(charset)
+        )
+    }
+
+    override suspend fun deserialize(charset: Charset, typeInfo: TypeInfo, content: ByteReadChannel): Any? {
+        return withContext(Dispatchers.IO) {
+            val reader = content.toInputStream().reader(charset)
+            objectmapper.readValue(reader, objectmapper.constructType(typeInfo.reifiedType))
+        }
+    }
+}
+
+/**
+ * Register Jackson converter into [ContentNegotiation] feature
+ */
+public fun Configuration.jackson(
+    contentType: ContentType = ContentType.Application.Json,
+    block: ObjectMapper.() -> Unit = {}
+) {
+    val mapper = ObjectMapper()
+    mapper.apply {
+        setDefaultPrettyPrinter(
+            DefaultPrettyPrinter().apply {
+                indentArraysWith(DefaultPrettyPrinter.FixedSpaceIndenter.instance)
+                indentObjectsWith(DefaultIndenter("  ", "\n"))
+            }
+        )
+    }
+    mapper.apply(block)
+    mapper.registerKotlinModule()
+    val converter = JacksonConverter(mapper)
+    register(contentType, converter)
+}
