+++ ktor-shared/ktor-shared-serialization-kotlinx/common/src/io/ktor/shared/serialization/kotlinx/SerializerLookup.kt
@@ -1,0 +1,66 @@
+/*
+ * Copyright 2014-2021 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+package io.ktor.shared.serialization.kotlinx
+
+import io.ktor.util.reflect.*
+import kotlinx.serialization.*
+import kotlinx.serialization.builtins.*
+import kotlinx.serialization.json.*
+import kotlinx.serialization.modules.*
+
+@OptIn(InternalSerializationApi::class, ExperimentalSerializationApi::class)
+internal fun serializerFromTypeInfo(
+    typeInfo: TypeInfo,
+    module: SerializersModule
+): KSerializer<*>? {
+    return module.getContextual(typeInfo.type)
+        ?: typeInfo.kotlinType?.let { module.serializerOrNull(it) }
+        ?: typeInfo.type.serializer()
+}
+
+@Suppress("UNCHECKED_CAST")
+internal fun guessSerializer(value: Any, module: SerializersModule): KSerializer<Any> = when (value) {
+    is JsonElement -> JsonElement.serializer()
+    is List<*> -> ListSerializer(value.elementSerializer(module))
+    is Array<*> -> value.firstOrNull()?.let { guessSerializer(it, module) } ?: ListSerializer(String.serializer())
+    is Set<*> -> SetSerializer(value.elementSerializer(module))
+    is Map<*, *> -> {
+        val keySerializer = value.keys.elementSerializer(module)
+        val valueSerializer = value.values.elementSerializer(module)
+        MapSerializer(keySerializer, valueSerializer)
+    }
+    else -> {
+        @OptIn(InternalSerializationApi::class, ExperimentalSerializationApi::class)
+        module.getContextual(value::class) ?: value::class.serializer()
+    }
+} as KSerializer<Any>
+
+@OptIn(ExperimentalSerializationApi::class)
+private fun Collection<*>.elementSerializer(module: SerializersModule): KSerializer<*> {
+    val serializers: List<KSerializer<*>> =
+        filterNotNull().map { guessSerializer(it, module) }.distinctBy { it.descriptor.serialName }
+
+    if (serializers.size > 1) {
+        error(
+            "Serializing collections of different element types is not yet supported. " +
+                "Selected serializers: ${serializers.map { it.descriptor.serialName }}"
+        )
+    }
+
+    val selected = serializers.singleOrNull() ?: String.serializer()
+
+    if (selected.descriptor.isNullable) {
+        return selected
+    }
+
+    @Suppress("UNCHECKED_CAST")
+    selected as KSerializer<Any>
+
+    if (any { it == null }) {
+        return selected.nullable
+    }
+
+    return selected
+}
