+++ ktor-shared/ktor-shared-serialization/build.gradle
@@ -1,0 +1,2 @@
+description = 'Serialization API for client and server'
+
