+++ settings.gradle
@@ -78,8 +78,6 @@ include ':ktor-features:ktor-mustache'
 include ':ktor-features:ktor-thymeleaf'
 include ':ktor-features:ktor-velocity'
 include ':ktor-features:ktor-pebble'
-include ':ktor-features:ktor-gson'
-include ':ktor-features:ktor-jackson'
 include ':ktor-features:ktor-metrics'
 include ':ktor-features:ktor-metrics-micrometer'
 include ':ktor-features:ktor-server-sessions'
@@ -90,7 +88,6 @@ include ':ktor-features:ktor-auth'
 include ':ktor-features:ktor-auth-ldap'
 include ':ktor-features:ktor-auth-jwt'
 include ':ktor-features:ktor-webjars'
-include ':ktor-features:ktor-serialization'
 include ':ktor-features'
 include ':ktor-http'
 include ':ktor-http:ktor-http-cio'
@@ -101,3 +98,8 @@ include ':ktor-network:ktor-network-tls'
 include ':ktor-network:ktor-network-tls:ktor-network-tls-certificates'
 include ':ktor-bom'
 include ':ktor-test-dispatcher'
+include ':ktor-shared'
+include ':ktor-shared:ktor-shared-serialization'
+include ':ktor-shared:ktor-shared-serialization-kotlinx'
+include ':ktor-shared:ktor-shared-serialization-gson'
+include ':ktor-shared:ktor-shared-serialization-jackson'
